test/api/openAI.test.ts
// 1. Successful Response Test
// Test Name: should return generated text for a valid prompt
// Purpose: Tests that the function returns the expected text when provided with a valid prompt.
// Method: Mock the OpenAI API response to return a predefined result for a specific prompt and verify that the function returns this result.
// 2. Handling API Errors
// Test Name: should throw an error when the OpenAI API returns an error
// Purpose: Ensures the function properly handles and relays errors from the OpenAI API, such as rate limits exceeded or invalid API keys.
// Method: Mock the OpenAI API to return an error response and verify that the function throws an error or rejects the promise.
// 3. Handling Network Issues
// Test Name: should handle network or connectivity issues gracefully
// Purpose: Tests the function's ability to handle exceptions related to network issues, such as timeouts or unreachable API endpoints.
// Method: Simulate a network failure and ensure the function handles it gracefully, either by retrying, throwing a specific error, or failing in a controlled manner.
// 4. Custom Settings Effect
// Test Name: should correctly apply custom settings to the API request
// Purpose: Verifies that any custom settings (e.g., model, temperature, max_tokens) are correctly applied to the API request.
// Method: Mock the API request and verify that the custom settings are included in the request payload.
// 5. Validation of Input
// Test Name: should validate input and throw an error for invalid prompts
// Purpose: Ensures the function validates input prompts and throws an error for invalid inputs before making an API call.
// Method: Call the function with various invalid inputs and expect errors.
// 6. Response Parsing
// Test Name: should correctly parse and format the API response
// Purpose: Tests that the function correctly parses the API response and formats it according to the expected output format, handling any variations in the API's response structure.
// Method: Mock the OpenAI API to return responses with different structures and verify that the function returns the correctly parsed text.
// 7. Rate Limit Handling
// Test Name: should handle rate limits correctly
// Purpose: Ensures that the function has a strategy for handling rate limits, either by queuing requests, delaying retries, or notifying the user.
// Method: Mock the API to simulate a rate limit response and test how the function responds.
// Mocking and Tools
// For these tests, you'll need to use mocking tools (like Jest's mocking functionalities if you're using Jest) to simulate OpenAI API responses. This approach allows you to test how your function behaves under different scenarios without making actual API calls, thus avoiding unnecessary costs and the unpredictability of live network conditions.

import axios from 'axios';
import { generateText } from '../../src/api/openAI';
jest.mock('axios');
const mockedAxios = axios as jest.Mocked<typeof axios>;

describe('openAI', () => {
  beforeEach(() => {
    // Clear all mocks before each test
    mockedAxios.post.mockClear();
  });

  it('should return generated text for a valid prompt', async () => {
    const fakeResponse = { data: { choices: [{ text: 'Hello, world!' }] } };
    mockedAxios.post.mockResolvedValue(fakeResponse);

    const prompt = 'Say hello';
    const response = await generateText(prompt);

    expect(response).toEqual('Hello, world!');
    expect(mockedAxios.post).toHaveBeenCalledWith(expect.any(String), expect.objectContaining({
      prompt: prompt
    }), expect.any(Object));
  });

  it('should throw an error when the OpenAI API returns an error', async () => {
    mockedAxios.post.mockRejectedValue(new Error('API error'));

    await expect(generateText('Say hello')).rejects.toThrow('API error');
  });

  // Add more tests here following the outlined scenarios

});
